move relay::v2 root symbols to relay::
and use them across the crate. Deprecate the old ones.
